Improved docs for a bunch of items
Improved docs for items from `array` up to `option` module
(no improvements in the root module itself yet).

Removed `__priv_transmute_ref`/`__priv_transmute_mut`,
they can be done purely with pointer casts and dereferencing.

+++ konst/src/iter/collect_const.rs
@@ -34,7 +34,7 @@ the [`iterator_dsl`] module.
 ```rust
 use konst::iter;
 
-const ARR: [usize; 8] = iter::collect_const!(usize =>
+const ARR: [u64; 8] = iter::collect_const!(u64 =>
     10..,
         filter(|n| *n % 2 == 0),
         skip(5),
